+++ app_server/api/v1/views/projects.py
@@ -4,8 +4,6 @@ from sqlalchemy.sql import text
 from sqlalchemy.exc import OperationalError
 import sqlalchemy.orm.collections as soc
 
-
-
 from api.v1.views import app_bp
 from db import main_storage
 from db.db_classes import User, Project, Task, Note
@@ -59,7 +57,6 @@ def user_project_route(user_id, project_id):
             return jsonify({})
 
 
-
 @app_bp.route('/<user_id>/projects/<project_id>/tasks',
              methods=['GET', 'POST'], strict_slashes=False)
 def user_project_tasks_route(user_id, project_id):
@@ -71,24 +68,23 @@ def user_project_tasks_route(user_id, project_id):
 
     if project is None or project.user_id != user.id:
         return jsonify({'error': 'project not found'})
-    match request.method:
-        case 'GET':
-            return jsonify([task.dict_repr() for task in project.tasks])
-        case 'POST':
-            try:
-                task_dict = request.get_json(silent=True)
-                # ensure correct user id and project id
-                task_dict['user_id'] = user_id
-                task_dict['project_id'] = project_id
-                task = Task(**task_dict)
-                main_storage.new(task)
-                main_storage.save()
-                return jsonify(task.dict_repr())
-            except OperationalError:
-                return jsonify({'error': 'json missing project parameters'})
-            except Exception as e:
-                print(f"[{e.__class__.__name__}]: {e}")
-                return jsonify({})
+    if request.method == 'GET':
+        return jsonify([task.dict_repr() for task in project.tasks])
+    elif request.method == 'POST':
+        try:
+            task_dict = request.get_json(silent=True)
+            # ensure correct user id and project id
+            task_dict['user_id'] = user_id
+            task_dict['project_id'] = project_id
+            task = Task(**task_dict)
+            main_storage.new(task)
+            main_storage.save()
+            return jsonify(task.dict_repr())
+        except OperationalError:
+            return jsonify({'error': 'json missing project parameters'})
+        except Exception as e:
+            print(f"[{e.__class__.__name__}]: {e}")
+            return jsonify({})
 
 
 @app_bp.route('/<user_id>/projects/<project_id>/tasks/<task_id>',
@@ -106,23 +102,21 @@ def user_project_task_route(user_id, project_id, task_id):
     if task not in project.tasks or task.user_id != user_id:
         return jsonify({'error': 'task not found'})
 
-    match request.method:
-        case 'GET':
-            return jsonify(task.dict_repr())
-        case 'PUT':
-            dic = request.get_json(silent=True)
-            ##### THIS LEVEL OF INPUT HANDLING NEEDS TO BE ADDED OVER ALL ROUTES
-            if type(dic) is not dict:
-                return jsonify({'error': 'invalid request'})
-            task.update(dic)
-            return jsonify(task.dict_repr())
-        case 'DELETE':
-            main_storage.delete(task)
-            main_storage.save()
-            if main_storage.get(Task, user_id) is not None:
-                return jsonify({'error': 'there was a problem deleting the instance'})
-            return jsonify({})
-
+    if request.method == 'GET':
+        return jsonify(task.dict_repr())
+    elif request.method == 'PUT':
+        dic = request.get_json(silent=True)
+        ##### THIS LEVEL OF INPUT HANDLING NEEDS TO BE ADDED OVER ALL ROUTES
+        if type(dic) is not dict:
+            return jsonify({'error': 'invalid request'})
+        task.update(dic)
+        return jsonify(task.dict_repr())
+    elif request.method == 'DELETE':
+        main_storage.delete(task)
+        main_storage.save()
+        if main_storage.get(Task, user_id) is not None:
+            return jsonify({'error': 'there was a problem deleting the instance'})
+        return jsonify({})
 
 
 @app_bp.route('/<user_id>/projects/<project_id>/notes',
@@ -136,24 +130,25 @@ def user_project_notes_route(user_id, project_id):
 
     if project is None or project.user_id != user.id:
         return jsonify({'error': 'project not found'})
-    match request.method:
-        case 'GET':
-            return jsonify([note.dict_repr() for note in project.notes])
-        case 'POST':
-            try:
-                note_dict = request.get_json(silent=True)
-                # ensure correct user id and project id
-                note_dict['user_id'] = user_id
-                note_dict['project_id'] = project_id
-                note = Note(**note_dict)
-                main_storage.new(note)
-                main_storage.save()
-                return jsonify(note.dict_repr())
-            except OperationalError:
-                return jsonify({'error': 'json missing project parameters'})
-            except Exception as e:
-                print(f"[{e.__class__.__name__}]: {e}")
-                return jsonify({})
+
+    if request.method == 'GET':
+        return jsonify([note.dict_repr() for note in project.notes])
+    elif request.method == 'POST':
+        try:
+            note_dict = request.get_json(silent=True)
+            # ensure correct user id and project id
+            note_dict['user_id'] = user_id
+            note_dict['project_id'] = project_id
+            note = Note(**note_dict)
+            main_storage.new(note)
+            main_storage.save()
+            return jsonify(note.dict_repr())
+        except OperationalError:
+            return jsonify({'error': 'json missing project parameters'})
+        except Exception as e:
+            print(f"[{e.__class__.__name__}]: {e}")
+            return jsonify({})
+
 
 @app_bp.route('/<user_id>/projects/<project_id>/notes/<note_id>',
              methods=['GET', 'PUT', 'DELETE'], strict_slashes=False)
@@ -170,18 +165,17 @@ def user_project_note_route(user_id, project_id, note_id):
     if note is None or note.user_id != user_id or note.project_id != project_id:
         return jsonify({'error': 'note not found'})
 
-    match request.method:
-        case 'GET':
+    if request.method == 'GET':
+        return jsonify(note.dict_repr())
+    elif request.method == 'PUT':
+        dic = request.get_json(silent=True)
+        if type(dic) is dict:
+            note.update(dic)
             return jsonify(note.dict_repr())
-        case 'PUT':
-            dic = request.get_json(silent=True)
-            if type(dic) is dict:
-                note.update(dic)
-                return jsonify(note.dict_repr())
-            return jsonify({})
-        case 'DELETE':
-            with main_storage.engine.connect() as con:
-                con.execute(text(f"DELETE FROM notes WHERE id = '{note_id}';"))
-                con.commit()
-            main_storage.save()
-            return jsonify({})
+        return jsonify({})
+    elif request.method == 'DELETE':
+        with main_storage.engine.connect() as con:
+            con.execute(text(f"DELETE FROM notes WHERE id = '{note_id}';"))
+            con.commit()
+        main_storage.save()
+        return jsonify({})
